2.ChatModel/3_Chatmodel_google.py
from langchain_google_genai import ChatGoogleGenerativeAI
from dotenv import load_dotenv

load_dotenv()

model =ChatGoogleGenerativeAI(model='gemini-3.1-flash-image-preview')
result= model.invoke("what is capital of india?")
print(result.content)